Update main code
Added a confirmation loop, so that the program will only go forward with the actual guessing game if the user confirms the proper amount of letters in their word. Next, I will add a way for the program to delete all words in its list that does not have the amount of letters the user inputted it.